opencdx-protector/src/main/java/cdx/opencdx/protector/service/impl/OpenCDXProtectorServiceImpl.java
/*
 * Copyright 2024 Safe Health Systems, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package cdx.opencdx.protector.service.impl;

import cdx.opencdx.commons.service.OpenCDXDocumentValidator;
import cdx.opencdx.grpc.neural.protector.*;
import cdx.opencdx.protector.service.OpenCDXProtectorService;
import io.micrometer.observation.annotation.Observed;
import lombok.extern.slf4j.Slf4j;
import org.bson.types.ObjectId;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

/**
 * Service for processing protector Requests
 */
@Slf4j
@Service
@Observed(name = "opencdx")
public class OpenCDXProtectorServiceImpl implements OpenCDXProtectorService {

    private static final String USERS = "users";

    private final OpenCDXDocumentValidator openCDXDocumentValidator;

    /**
     * Constructor for OpenCDXProtectorServiceImpl.
     *
     * @param openCDXDocumentValidator Document validator
     */
    @Autowired
    public OpenCDXProtectorServiceImpl(OpenCDXDocumentValidator openCDXDocumentValidator) {
        this.openCDXDocumentValidator = openCDXDocumentValidator;
    }

    @Override
    public SecurityResponse detectAnomalies(AnomalyDetectionDataRequest request) {
        this.openCDXDocumentValidator.validateDocumentOrThrow(
                USERS, new ObjectId(request.getAnomalyDetectionData().getUserId()));

        // Placeholder implementation for detecting anomalies
        // Actual implementation should analyze the request and provide a SecurityResponse
        return SecurityResponse.newBuilder()
                .setEncounterId(request.getAnomalyDetectionData().getEncounterId())
                .setResponse("SecurityResponse [detectAnomalies]")
                .build();
    }

    @Override
    public SecurityResponse enforceAuthorizationControl(AuthorizationControlDataRequest request) {
        this.openCDXDocumentValidator.validateDocumentOrThrow(
                USERS, new ObjectId(request.getAuthorizationControlData().getUserId()));

        // Placeholder implementation for enforcing authorization control
        // Actual implementation should analyze the request and provide a SecurityResponse
        return SecurityResponse.newBuilder()
                .setEncounterId(request.getAuthorizationControlData().getEncounterId())
                .setResponse("SecurityResponse [enforceAuthorizationControl]")
                .build();
    }

    @Override
    public SecurityResponse protectPrivacy(PrivacyProtectionDataRequest request) {

        // Placeholder implementation for privacy protection
        // Actual implementation should analyze the request and provide a SecurityResponse
        return SecurityResponse.newBuilder()
                .setEncounterId(request.getPrivacyProtectionData().getEncounterId())
                .setResponse("SecurityResponse [protectPrivacy]")
                .build();
    }

    @Override
    public SecurityResponse monitorRealTimeActivity(RealTimeMonitoringDataRequest request) {
        this.openCDXDocumentValidator.validateDocumentOrThrow(
                USERS, new ObjectId(request.getRealTimeMonitoringData().getMonitoredEntity()));

        // Placeholder implementation for monitoring real-time activity
        // Actual implementation should analyze the request and provide a SecurityResponse
        return SecurityResponse.newBuilder()
                .setEncounterId(request.getRealTimeMonitoringData().getEncounterId())
                .setResponse("SecurityResponse [monitorRealTimeActivity]")
                .build();
    }

    @Override
    public SecurityResponse analyzeUserBehavior(UserBehaviorAnalysisDataRequest request) {
        this.openCDXDocumentValidator.validateDocumentOrThrow(
                USERS, new ObjectId(request.getUserBehaviorAnalysisData().getUserId()));

        // Placeholder implementation for analyzing user behavior
        // Actual implementation should analyze the request and provide a SecurityResponse
        return SecurityResponse.newBuilder()
                .setEncounterId(request.getUserBehaviorAnalysisData().getEncounterId())
                .setResponse("SecurityResponse [analyzeUserBehavior]")
                .build();
    }
}
